+ Small changes to clean up warnings that appeared when we started   using gcc-4.7.1 for regression tests.

+++ src/parser/Abstract_Class_Parser.i.hh
@@ -53,7 +53,7 @@ register_child(string const &keyword,
     std::strcpy(cptr, keyword.c_str());
     keys_.push_back(cptr);
 
-    unsigned const N = map_.size();
+    int const N = static_cast<int>(map_.size());
     map_.push_back(parse_function);
     
     Keyword key = {cptr,
